+++ tests/backends/ldap/test_utils.py
@@ -280,20 +280,20 @@ def test_ldap_cannot_create_groups(testclient, configuration, backend):
 
 
 def test_login_placeholder(testclient):
-    testclient.app.config["BACKENDS"]["LDAP"]["USER_FILTER"] = "(uid={login})"
+    testclient.app.config["BACKENDS"]["LDAP"]["USER_FILTER"] = "(uid={{ login }})"
     placeholder = testclient.get("/login").form["login"].attrs["placeholder"]
     assert placeholder == "jdoe"
 
-    testclient.app.config["BACKENDS"]["LDAP"]["USER_FILTER"] = "(cn={login})"
+    testclient.app.config["BACKENDS"]["LDAP"]["USER_FILTER"] = "(cn={{ login }})"
     placeholder = testclient.get("/login").form["login"].attrs["placeholder"]
     assert placeholder == "John Doe"
 
-    testclient.app.config["BACKENDS"]["LDAP"]["USER_FILTER"] = "(mail={login})"
+    testclient.app.config["BACKENDS"]["LDAP"]["USER_FILTER"] = "(mail={{ login }})"
     placeholder = testclient.get("/login").form["login"].attrs["placeholder"]
     assert placeholder == "john@doe.com"
 
     testclient.app.config["BACKENDS"]["LDAP"][
         "USER_FILTER"
-    ] = "(|(uid={login})(mail={login}))"
+    ] = "(|(uid={{ login }})(mail={{ login }}))"
     placeholder = testclient.get("/login").form["login"].attrs["placeholder"]
     assert placeholder == "jdoe or john@doe.com"